// packages/api-client/src/generated/model/listDocumentsSortDir.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Document Intake API
 * Intake queue for documents captured by the Document Intake PCF control.
 * OpenAPI spec version: 1.0.0
 */

export type ListDocumentsSortDir = typeof ListDocumentsSortDir[keyof typeof ListDocumentsSortDir];


export const ListDocumentsSortDir = {
  asc: 'asc',
  desc: 'desc',
} as const;